Update EventsHandlerV2.cs
- Fixed a bug that caused the mod to crash if an event uses a float value inside at an "change_employee_stress" or "change_employee_satisfaction" effect

Added to Steam version 2.3.1

+++ src/EventsHandlerV2.cs
@@ -1158,14 +1158,14 @@ namespace RandomEvents
 				{
 					if (actor != null)
 					{
-						actor.employee.JobSatisfaction += int.Parse(EffectValue);
+						actor.employee.JobSatisfaction += float.Parse(EffectValue);
 					}
 				}
 				else if (Effect == "change_employee_stress")
 				{
 					if (actor != null)
 					{
-						actor.employee.Stress += int.Parse(EffectValue);
+						actor.employee.Stress += float.Parse(EffectValue);
 					}
 				}
 				else if (Effect == "change_bugs")
